changes in waypoint handling and db

+++ server/outdoortribeDB.sql
@@ -292,7 +292,7 @@ CREATE TABLE `waypoints` (
   `id` int NOT NULL,
   `post_id` int NOT NULL,
   `coordinates` varchar(255) CHARACTER SET utf8mb4 COLLATE utf8mb4_unicode_ci DEFAULT NULL,
-  `km` int NOT NULL,
+  `km` float NOT NULL,
   `title` varchar(255) CHARACTER SET utf8mb4 COLLATE utf8mb4_unicode_ci NOT NULL,
   `description` text CHARACTER SET utf8mb4 COLLATE utf8mb4_unicode_ci
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_unicode_ci;
